// app/verticals/service-jobs/shared/apis/create-service-job-action.ts
export const CreateServiceJobActionCommitIndeterminateProblemSchema = makeProblemDetailsSchema(
  'CreateServiceJobActionCommitIndeterminateProblem',
  503,
  {
    invocationId: Schema.String,
    resolution: Schema.Literal('RESOLVE_COMMIT'),
    retryCommand: Schema.Literal(false),
  },
);